Split top level porting details from main() for library build Move porting details to unixport.[c,h]. No change in code. mkrules.mk library rule assumes same objects as main executable so I created a separate group LIBOBJ. There is probably a more elegant way to do this.
